doc-source/requirements.txt
git+https://github.com/sphinx-toolbox/html-section.git
autodocsumm>=0.2.0
default-values>=0.4.2
extras-require>=0.2.0
furo>=2020.11.19b18
seed-intersphinx-mapping>=0.3.1
sphinx-copybutton>=0.2.12
sphinx-debuginfo>=0.1.0
sphinx-notfound-page>=0.5
sphinx-prompt>=1.1.0
sphinx-pyproject>=0.1.0
sphinx-tabs>=1.1.13
sphinx-toolbox>=2.2.0
sphinxcontrib-httpdomain>=1.7.0
sphinxemoji>=0.1.6
toctree-plus>=0.1.0
